+++ Student1.ParentPortal.Data/Models/EdFi31/StudentRepository.cs
@@ -315,7 +315,7 @@ namespace Student1.ParentPortal.Data.Models.EdFi31
                               {
                                   Title = a.AssessmentTitle,
                                   Identifier = a.AssessmentIdentifier,
-                                  MaxRawScore = a.MaxRawScore,
+                                  MaxRawScore = (decimal)a.MaxRawScore,
                                   AdministrationDate = sa.AdministrationDate,
                                   Result = sasr.Result,
                                   PerformanceLevelMet = plt.ShortDescription,
@@ -1273,7 +1273,7 @@ namespace Student1.ParentPortal.Data.Models.EdFi31
                                   //Version = a.Version,
                                   Title = a.AssessmentTitle,
                                   Identifier = a.AssessmentIdentifier,
-                                  MaxRawScore = a.MaxRawScore,
+                                  MaxRawScore = (decimal)a.MaxRawScore,
                                   AdministrationDate = sa.AdministrationDate,
                                   Result = sasr.Result,
                                   //ReportingMethodCodeValue = sasr.AssessmentReportingMethodType.CodeValue
@@ -1300,7 +1300,7 @@ namespace Student1.ParentPortal.Data.Models.EdFi31
                                   //Version = a.Version,
                                   Title = a.AssessmentTitle,
                                   Identifier = a.AssessmentIdentifier,
-                                  MaxRawScore = a.MaxRawScore,
+                                  MaxRawScore = (decimal)a.MaxRawScore,
                                   AdministrationDate = sa.AdministrationDate,
                                   PerformanceLevelMet = d.CodeValue,
                                   //ReportingMethodCodeValue = sapl.AssessmentReportingMethodType.CodeValue
@@ -1327,7 +1327,7 @@ namespace Student1.ParentPortal.Data.Models.EdFi31
                                   //Version = a.Version,
                                   Title = a.AssessmentTitle,
                                   Identifier = a.AssessmentIdentifier,
-                                  MaxRawScore = a.MaxRawScore,
+                                  MaxRawScore = (decimal)a.MaxRawScore,
                                   AdministrationDate = sa.AdministrationDate,
                                   Result = sasr.Result,
                                   //ReportingMethodCodeValue = sasr.AssessmentReportingMethodType.CodeValue,
